pyproject.toml
[project]
name = "synthetic-data-agent"
version = "0.1.0"
description = "Production-grade synthetic data generation system using Google ADK"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    # ADK + A2A
    "google-adk[a2a]>=1.0.0",
    "a2a>=0.2.0",

    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",

    # Config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Databricks
    "databricks-sdk>=0.28.0",
    "databricks-connect>=14.3.0",

    # Synthetic data ML
    "sdv>=1.15.0",
    "ctgan>=0.10.0",
    "torch>=2.3.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # PII detection
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "faker>=25.0.0",
    "spacy>=3.7.0",

    # Statistical profiling
    "ydata-profiling>=4.8.0",

    # Quality validation
    "sdmetrics>=0.16.0",

    # Storage + caching
    "redis>=5.0.0",
    "joblib>=1.4.0",

    # Observability
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-api>=1.25.0",
    "structlog>=24.2.0",

    # Utilities
    "networkx>=3.3",
    "cachetools>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-redis>=4.6.0",
    "pandas-stubs>=2.2.0",
    "syrupy>=4.0",
]

[project.scripts]
synthetic-data-agent = "synthetic_data_agent.server.main:main"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "PT", "RUF"]
ignore = [
    "ANN101", "ANN102",
    "S101",    # assert in tests
    "S603",    # subprocess
    "ANN401",  # Any in specific places
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S", "N"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
python_version = "3.12"

[[tool.mypy.overrides]]
module = [
    "sdv.*",
    "ctgan.*",
    "presidio_analyzer.*",
    "presidio_anonymizer.*",
    "ydata_profiling.*",
    "sdmetrics.*",
    "databricks.*",
    "networkx.*",
    "faker.*",
    "spacy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/synthetic_data_agent"]
